Break out point tree headers and start better disjoint
